+++ src/Commands/ListenerMakeCommand.php
@@ -28,17 +28,6 @@ class ListenerMakeCommand extends GeneratorCommand
      */
     protected $description = 'Create a new event listener class for the specified module';
 
-    public function handle()
-    {
-        if (! $this->option('event')) {
-            $this->error('The --event option is necessary');
-
-            return;
-        }
-
-        parent::handle();
-    }
-
     /**
      * Get the console command arguments.
      *
@@ -60,7 +49,8 @@ class ListenerMakeCommand extends GeneratorCommand
     protected function getOptions()
     {
         return [
-            ['event', null, InputOption::VALUE_REQUIRED, 'Event name this is listening to', null],
+            ['event', 'e', InputOption::VALUE_OPTIONAL, 'The event class being listened for.'],
+            ['queued', null, InputOption::VALUE_NONE, 'Indicates the event listener should be queued.'],
         ];
     }
 
@@ -68,10 +58,11 @@ class ListenerMakeCommand extends GeneratorCommand
     {
         $module = $this->laravel['modules']->findOrFail($this->getModuleName());
 
-        return (new Stub('/listener.stub', [
+        return (new Stub($this->getStubName(), [
             'NAMESPACE' => $this->getNamespace($module),
-            "EVENTNAME" => $this->getEventName($module),
-            "CLASS"     => $this->getClass(),
+            'EVENTNAME' => $this->getEventName($module),
+            'SHORTEVENTNAME' => $this->option('event'),
+            'CLASS' => $this->getClass(),
         ]))->render();
     }
 
@@ -102,5 +93,25 @@ class ListenerMakeCommand extends GeneratorCommand
     protected function getFileName()
     {
         return studly_case($this->argument('name'));
+    }
+
+    /**
+     * @return string
+     */
+    protected function getStubName(): string
+    {
+        if ($this->option('queued')) {
+            if ($this->option('event')) {
+                return '/listener-queued.stub';
+            }
+
+            return '/listener-queued-duck.stub';
+        }
+
+        if ($this->option('event')) {
+            return '/listener.stub';
+        }
+
+        return '/listener-duck.stub';
     }
 }