+++ tests/Commands/__snapshots__/ListenerMakeCommandTest__it_generated_correct_sync_duck_event_with_content__1.php
@@ -2,7 +2,6 @@
 
 namespace Modules\\Blog\\Listeners;
 
-use Modules\\Blog\\Events\\UserWasCreated;
 use Illuminate\\Queue\\InteractsWithQueue;
 use Illuminate\\Contracts\\Queue\\ShouldQueue;
 
@@ -21,10 +20,10 @@ class NotifyUsersOfANewPost
     /**
      * Handle the event.
      *
-     * @param \\Modules\\Blog\\Events\\UserWasCreated $event
+     * @param  object  $event
      * @return void
      */
-    public function handle(\\Modules\\Blog\\Events\\UserWasCreated $event)
+    public function handle($event)
     {
         //
     }